==> 2019-email-survey-analysis/01 Data Exploration.Rmd <==
---
title: "01 Data exploration"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Setup

```{r setup}
library(tidyverse)
library(lubridate)
library(readxl)
library(wranglR)
```

Some data validation functions.

```{r}
# Check for duplicate rows, based on some single-column unique identifier
check_dupes <- function(dataframe, unique_id_colname) {
  print(
    paste(
      'Duplicates?'
      , nrow(dataframe) != dataframe %>% select_(unique_id_colname) %>% unique() %>% nrow()
    )
  )
}
```

# Data

Load the raw iModules data.

```{r}
# Email sends
sends <- lapply(
  Sys.glob('data_imodules/*.xls*')
  , read_xlsx
  , sheet = 1
)
dat_sends <- data.frame()
for(i in 1:length(sends)) {
  dat_sends <- dat_sends %>% rbind(sends[[i]])
}

# Email bounces
bounces <- lapply(
  Sys.glob('data_imodules/*.xls*')
  # Error handling for missing sheets
  , function(x) {
      return(
        tryCatch(
          read_xlsx(x, sheet = 'Bounces')
          , error = function(y) {NULL}
        )
      )
    }
)
dat_bounces <- data.frame()
for(i in 1:length(bounces)) {
  dat_bounces <- dat_bounces %>% rbind(bounces[[i]])
}

# Email opens
opens <- lapply(
  Sys.glob('data_imodules/*.xls*')
  , read_xlsx
  , sheet = 'Opens'
)
dat_opens <- data.frame()
for(i in 1:length(opens)) {
  dat_opens <- dat_opens %>% rbind(opens[[i]])
}

# Email clicks
clicks <- lapply(
  Sys.glob('data_imodules/*.xls*')
  , read_xlsx
  , sheet = 'Clicks'
)
dat_clicks <- data.frame()
for(i in 1:length(clicks)) {
  dat_clicks <- dat_clicks %>% rbind(clicks[[i]])
}

# Email-ID lookup table
ids <- read_xlsx(path = 'data/alumni emails.xlsx', sheet = 'Select email') %>%
  mutate(ID_NUMBER = as.numeric(ID_NUMBER))
```

Some emails are not associated with CAT IDs; look them up where available.

```{r}
# Append to dat_sends
dat_sends <- dat_sends %>%
  left_join(ids, by = c('recipientEmailAddress' = 'EMAIL_ADDRESS')) %>%
  mutate(
    id_final = case_when(
      !is.na(recipientConstituentId) ~ recipientConstituentId
      , TRUE ~ ID_NUMBER
    )
    , deliveredTimestamp = ymd_hms(deliveredTimestamp)
    , dateAdded = ymd_hms(dateAdded)
  )

# Append to dat_bounces
dat_bounces <- dat_bounces %>%
  mutate(recipientConstituentId = as.numeric(recipientConstituentId)) %>%
  left_join(ids, by = c('recipientEmailAddress' = 'EMAIL_ADDRESS')) %>%
  mutate(
    id_final = case_when(
      !is.na(recipientConstituentId) ~ recipientConstituentId
      , TRUE ~ ID_NUMBER
    )
    , bouncedTimestamp = ymd_hms(bouncedTimestamp)
  )

# Append to dat_opens
dat_opens <- dat_opens %>%
  left_join(ids, by = c('recipientEmailAddress' = 'EMAIL_ADDRESS')) %>%
  mutate(
    id_final = case_when(
      !is.na(recipientConstituentId) ~ recipientConstituentId
      , TRUE ~ ID_NUMBER
    )
    , openedTimestamp = ymd_hms(openedTimestamp)
    , lastActionTime = ymd_hms(lastActionTime)
  )

# Append to dat_clicks
dat_clicks <- dat_clicks %>%
  left_join(ids, by = c('emailAddress' = 'EMAIL_ADDRESS')) %>%
  mutate(
    id_final = case_when(
      !is.na(constituentId) ~ constituentId
      , TRUE ~ ID_NUMBER
    )
  )
```

# iModules data cleanup

Check for missing IDs.

```{r}
dat_sends %>% select(recipientConstituentId, ID_NUMBER, id_final) %>% summary()
```

```{r}
dat_bounces %>% select(recipientConstituentId, ID_NUMBER, id_final) %>% summary()
```


```{r}
dat_opens %>% select(recipientConstituentId, ID_NUMBER, id_final) %>% summary()
```

```{r}
dat_clicks %>% select(constituentId, ID_NUMBER, id_final) %>% summary()
```

Check whether the missing IDs are actually alumni.

```{r}
dat_sends %>%
  filter(is.na(id_final)) %>%
  mutate(ksm = str_detect(recipientEmailAddress, pattern = '.*kellogg.*')) %>%
  select(ksm, id_final, recipientEmailAddress) %>%
  distinct() %>%
  group_by(ksm) %>%
  summarise(count = length(id_final))
```

Looks like I need to do some manual lookups, but it's not bad.

```{r}
# Write records missing IDs to file
writexl::write_xlsx(
  x = list(
    sends = dat_sends %>% filter(is.na(id_final))
    , bounces = dat_bounces %>% filter(is.na(id_final))
    , opens = dat_opens %>% filter(is.na(id_final))
    , clicks = dat_clicks %>% filter(is.na(id_final))
  )
  , path = 'data/missing ids.xlsx'
)
```

```{r}
# Pull ID lookups
id_lookups <- read_xlsx(path = 'data/missing ids lookups.xlsx', sheet = 1)

# Fill in missing IDs
fill_ids <- function(data, join_col) {
  data %>%
    left_join(
      id_lookups %>% select(recipientEmailAddress, ID)
      , by = setNames('recipientEmailAddress', join_col)
    ) %>%
    mutate(
      id_final = case_when(
        is.na(id_final) ~ ID
        , TRUE ~ id_final
      )
    ) %>%
    select(recipientId:id_final) %>%
    return()
}

dat_sends <- dat_sends %>% fill_ids('recipientEmailAddress')
dat_bounces <- dat_bounces %>% fill_ids('recipientEmailAddress')
dat_opens <- dat_opens %>% fill_ids('recipientEmailAddress')
dat_clicks <- dat_clicks %>% fill_ids('emailAddress')
```

```{r}
data.frame(
  sends = is.na(dat_sends$id_final) %>% sum()
  , bounces = is.na(dat_bounces$id_final) %>% sum()
  , opens = is.na(dat_opens$id_final) %>% sum()
  , clicks = is.na(dat_clicks$id_final) %>% sum()
)
```

```{r}
data.frame(
  sends = dat_sends %>%
    filter(!str_detect(str_to_lower(recipientEmailAddress), '.*@kellogg.northwestern.*')) %>%
    select(id_final) %>% is.na() %>% sum()
  , bounces = dat_bounces %>%
    filter(!str_detect(str_to_lower(recipientEmailAddress), '.*@kellogg.northwestern.*')) %>%
    select(id_final) %>% is.na() %>% sum()
  , opens = dat_opens %>%
    filter(!str_detect(str_to_lower(recipientEmailAddress), '.*@kellogg.northwestern.*')) %>%
    select(id_final) %>% is.na() %>% sum()
  , clicks = dat_clicks %>%
    filter(!str_detect(str_to_lower(emailAddress), '.*@kellogg.northwestern.*')) %>%
    select(id_final) %>% is.na() %>% sum()
)
```

The only remaining NAs are associated with KSM email addresses or the survey sender.

# Class data sheet

```{r}
# Load data
dat_class <- read_xlsx(path = 'data/Class data.xlsx', sheet = 1) %>%
  mutate(ID_NUMBER = as.numeric(ID_NUMBER))
```

```{r}
dat_class  %>% left_join(
  dat_sends %>% select(id_final) %>% mutate(a = 1) %>% unique()
  , by = c('ID_NUMBER' = 'id_final')
) %>% mutate(id = !is.na(a)) %>%
  select(id) %>% summary()
```

Almost the entire class was emailable which surprises me.

```{r}
check_dupes(dat_class, 'ID_NUMBER')
```

# Email data check

Confirm that entities have at most one associated preferred email.

```{r}
check_dupes(
  dat_class %>%
    left_join(
      ids %>% filter(PREFERRED_IND == 'Y')
      , by = c('ID_NUMBER' = 'ID_NUMBER')
    )
  , 'ID_NUMBER'
)
```

# Alumni survey data

```{r}
# Load data
dat_survey_alum <- read_xlsx(path = 'data/20191126_NPS Results.xlsx', sheet = 1)

emplids <- read_xlsx(path = 'data/ID to EMPLID conversion.xlsx', sheet = 1) %>%
  mutate(
    ID_NUMBER = as.numeric(CATRACKS_ID)
    , EMPLID = as.numeric(EMPLID)
  ) %>% select(-CATRACKS_ID)

# Append id_numbers
dat_survey_alum <- dat_survey_alum %>%
  left_join(emplids, by = c('EmplID' = 'EMPLID'))

# Network survey
dat_survey_network <- read_xlsx(path = 'data/FY 18 Health of Network 9.14.2018 (Newest Verison).xlsx', sheet = 1, guess_max = 1E5) %>%
  rename(ID_NUMBER = `Catracks ID`)
```

```{r}
check_dupes(dat_survey_alum, 'ID_NUMBER')
```

```{r}
check_dupes(dat_survey_network, 'ID_NUMBER')
```

No duplicate records for either data frame.

```{r}
dat_survey_alum %>% mutate(id = is.na(ID_NUMBER)) %>% select(id) %>% summary()
```

That's great -- one id_number was not found. Check how many survey respondents received an email.

```{r}
dat_survey_alum %>%
  left_join(dat_sends, by = c('ID_NUMBER' = 'id_final')) %>%
  mutate(sent_email = !is.na(emailName)) %>%
  select(EmplID, sent_email) %>%
  distinct() %>%
  select(sent_email) %>%
  summary()
```

Looks good.

```{r}
dat_survey_network %>%
  left_join(dat_sends, by = c('ID_NUMBER' = 'id_final')) %>%
  mutate(sent_email = !is.na(emailName)) %>%
  select(ID_NUMBER, sent_email) %>%
  distinct() %>%
  select(sent_email) %>%
  summary()
```

Fewer responses than I'd like but still looks like useful data.

# Data file creation

The level of analysis should be an individual (as opposed to an email or a recipient). That suggests that `dat_class` is the base table, and other datapoints need to be aggregated to the individual level.

```{r, warning = FALSE}
# Aggregate data frame
dat_agg <- dat_class %>%
  # Append current email info
  left_join(
      ids %>% filter(PREFERRED_IND == 'Y')
      , by = c('ID_NUMBER' = 'ID_NUMBER')
  ) %>%
  # Aggregate all email data to the id_number level
  # Sends
  left_join(
    dat_sends %>% group_by(id_final) %>% summarise(
      send_class_year = max(recipientClassYear)
      , send_email = max(emailName)
      , send_email_used = max(recipientEmailAddress)
      , send_email_status = max(SHORT_DESC)
      , send_email_status_change_dt = max(STATUS_CHANGE_DATE)
      , send_email_dt_mod = max(DATE_MODIFIED)
      , last_delivered_dt = max(deliveredTimestamp)
      , n_delivered = length(deliveredTimestamp)
    )
    , by = c('ID_NUMBER' = 'id_final')
  ) %>%
  # Bounces
  left_join(
    dat_bounces %>% group_by(id_final) %>% summarise(
      last_bounced_dt = max(bouncedTimestamp)
      , n_bounced = length(bouncedTimestamp)
      , bounce_reason = max(bounceReason)
    )
    , by = c('ID_NUMBER' = 'id_final')
  ) %>%
  # Opens
  left_join(
    dat_opens %>% group_by(id_final) %>% summarise(
      last_opened_dt = max(openedTimestamp)
      , n_opened = length(openedTimestamp)
      , last_click_dt = max(lastActionTime)
    )
    , by = c('ID_NUMBER' = 'id_final')
  ) %>%
  # Clicks
  left_join(
    dat_clicks %>% group_by(id_final) %>%
      summarise(
        n_clicks = length(clickDate)
      )
    , by = c('ID_NUMBER' = 'id_final')
  ) %>%
  # Replace NA with 0 in count columns
  mutate_if(is.integer, replace_na, 0) %>%
  # Useful derived fields
  mutate(
    # Derived sent - delivered or bounced
    n_sent = n_delivered + n_bounced
    # Derived opens - open or click
    , n_opened_or_clicked = n_opened + n_clicks
  ) %>%
  # Alumni survey
  left_join(
    dat_survey_alum %>% select(ID_NUMBER, SurveyType:NPSCategory)
    , by = c('ID_NUMBER' = 'ID_NUMBER')
  ) %>%
  # Network survey
  left_join(
    dat_survey_network %>% select(ID_NUMBER, NPS_MBA_prog, Ethnicity, Income, NumberOrgs, StudentExp:Survey)
    , by = c('ID_NUMBER' = 'ID_NUMBER')
  ) %>%
  # Column name cleanup
  rename(
    EMAIL_STATUS = SHORT_DESC
  ) %>%
  # Fix data types
  mutate(
    RECORD_STATUS_CODE = as.factor(RECORD_STATUS_CODE)
    , FIRST_KSM_YEAR = as.numeric(FIRST_KSM_YEAR)
    , FIRST_MASTERS_YEAR = as.numeric(FIRST_MASTERS_YEAR)
    , PROGRAM = as.factor(PROGRAM)
    , PROGRAM_GROUP = as.factor(PROGRAM_GROUP) %>%
        fct_recode('Full-Time' = 'FT', 'EMBA' = 'EMP')
    , SURVEY = as.factor(SURVEY)
    , GRADS = as.factor(GRADS)
    , EMAIL_STATUS_CODE = as.factor(EMAIL_STATUS_CODE)
    , EMAIL_STATUS = as.factor(EMAIL_STATUS)
    , PREFERRED_IND = as.factor(PREFERRED_IND)
    , send_email_status = as.factor(send_email_status)
    , SurveyType = as.factor(SurveyType)
    , NPSCategory = as.factor(NPSCategory)
    , Ethnicity = as.factor(Ethnicity)
    , NPS_MBA_prog = as.integer(NPS_MBA_prog)
    , NumberOrgs = as.integer(NumberOrgs) %>% replace_na(0)
    , AlumNetworkResponse = as.factor(AlumNetworkResponse)
  ) %>%
  # Useful flags
  mutate(
    same_email = {str_trim(EMAIL_ADDRESS) %>% str_to_lower() == str_trim(send_email_used) %>% str_to_lower()} %>%
      replace_na(FALSE)
    , survey_nps_pct = NPS/NPSScale
  )
```

Determine which email group each alum belongs to.

```{r, rows.print = 100}
dat_agg %>%
  group_by(send_email, FIRST_MASTERS_YEAR, PROGRAM_GROUP) %>%
  summarize(n = length(ID_NUMBER))
```

```{r}
dat_agg <- dat_agg %>% mutate(
  email_group = case_when(
    is.na(send_email) ~
      'None'
    , send_email == 'RankingSurvey_2019EconomistWhichMBA_2018_2Y_04.16.19'
      ~ 'Economist 2018 FT'
    , send_email == 'RankingSurvey_Businessweek_2Y_2013_Reminder_07.11.19'
      ~ 'Businessweek 2013 FT'
    , send_email == 'RankingSurvey_BusinessWeek_FT2013_04.25.19'
      ~ 'Businessweek 2013 FT'
    , send_email == 'RankingSurvey_FT_2016EMBA_06.05.19'
      ~ 'Financial Times 2016 EMBA'
    , send_email == 'RankingSurvey_FT_EMBA(2016)_2ndReminder_07.23.19'
      ~ 'Financial Times 2016 EMBA'
    , send_email == 'RankingSurvey_FT_2016MBAs_09.05.19'
      ~ 'Financial Times 2016 FT'
  ) %>% as.factor()
)
```

```{r, rows.print = 100}
dat_agg %>%
  group_by(email_group, send_email, FIRST_MASTERS_YEAR, PROGRAM_GROUP) %>%
  summarize(n = length(ID_NUMBER))
```


# Output file

```{r}
save(dat_agg, file = 'data/dat_agg.Rdata')
```
