==> 2019-email-survey-analysis/02 Email statistics.Rmd <==
---
title: "Email statistics"
output:
  html_notebook:
    code_folding: hide
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

<style>
#header .btn-group {
    display: none;
}
</style>

```{r setup, echo = FALSE, message = FALSE, warning = FALSE}
# Libraries
library(tidyverse)
library(lubridate)
library(readxl)
library(wranglR)
library(knitr)
library(splines)
library(car)

# Parameters -- edit if needed
data_as_of <- ymd('20191202')

# Set to FALSE to hide all code chunks when rendering the document
opts_chunk$set(echo = FALSE)
```
```{r}
# Load data from last time
load('data/dat_agg.Rdata')
```


# Background and definitions

```{r, warning = FALSE}
# Email summary data from iModules
dat_summary_table <- read_xlsx(
  'data/Ranking Survey Email Metrics 18-19.xlsx'
  , sheet = 2
  , col_types = c('text', 'text', 'text', 'text', 'date', 'text', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'text')
) %>%
  filter(!is.na(`Email Type`) & Publication != 'Publication') %>%
  # Better column names
  rename(
    Drilldown = `Drill Down Data Available?`
  ) %>%
  # Format percents
  mutate(
    `Open Rate` = `Open Rate` %>% round(2) %>% scales::percent()
    , CTR = CTR %>% round(2) %>% scales::percent()
  )
```

iModules is the system used to send mass emails to Northwestern constituents, including alumni. Detailed email recipient, bounce, open, and click data is available for 480 days after an email is sent; after that time, only summary statistics are available.

The metrics are defined as follows:

* **Delivered** = 200 HTTP status code response from server, e.g. the request was fulfilled (*delivered does not imply inbox placement*)
* **Bounces** = email addresses that could not be successfully delivered to the recipient's inbox for some reason
* **Open** = tracking pixel loaded (*implying that image blocking or plain-text email clients leads to underreporting of opens*)
* **Click** = number of times any recipient clicks on any link within the email where tracking is active (*excluding the unsubscribe link*)

Summary statistics for each publication appear below. Looking at the email send dates, detailed information is available for just `r dat_summary_table %>% filter(Drilldown == 'Y') %>% nrow() %>% I()` of the `r dat_summary_table %>% nrow() %>% I()` emails sent (`Drilldown == 'Y'`).

# iModules summary statistics

## Businessweek

```{r}
dat_summary_table %>%
  filter(Publication == 'Businessweek') %>%
  select(-Publication, -`Time Sent`) %>%
  kable()
```

## Economist

```{r}
dat_summary_table %>%
  filter(Publication == 'Economist') %>%
  select(-Publication, -`Time Sent`) %>%
  kable()
```

## Financial Times

```{r}
dat_summary_table %>%
  filter(Publication == 'Financial Times') %>%
  select(-Publication, -`Time Sent`) %>%
  kable()
```

## Forbes

```{r}
dat_summary_table %>%
  filter(Publication == 'Forbes') %>%
  select(-Publication, -`Time Sent`) %>%
  kable()
```

Forbes email metrics are not available, as the emails were sent through Outlook rather than iModules (recipients had personalized links).

Overall, the bounce rates are surprisingly low, and the open/click rates quite high.

# Computed summary statistics

As a point of comparison, I pulled a full class list from the database. Now, the unit of analysis is individuals rather than emails. Individuals are categorized by the set of emails received, e.g. Economist 2018 FT is full-time 2018 graduates emailed about the Economist.

If an individual opened either an initial email or a reminder, I counted that as an open; the same goes for clicks. Note that raw counts will be different, as non-alumni recipients (such as KSM staff) are filtered out, unlike in the summary statistics above.

```{r}
# Summary table function
email_summary_table <- function(data, table_output = TRUE) {
  pct_format <- function(x, bool) {
    ifelse(bool, scales::percent(x), x)
  }
  data %>%
    summarize(
      n = length(ID_NUMBER)
      , sent = sum(ifelse(n_sent > 0, 1, 0))
      , bounced = sum(ifelse(n_bounced > 0, 1, 0))
      , opened = sum(ifelse(n_opened_or_clicked > 0, 1, 0))
      , clicked = sum(ifelse(n_clicks > 0, 1, 0))
    ) %>% mutate(
      pcf = table_output
      , sent_pct = pct_format(sent / n, pcf)
      , open_pct = pct_format(opened / sent, pcf)
      , clicked_pct = pct_format(clicked / (opened + 1E-100), pcf)
    ) %>%
    select(-pcf) %>%
    return()
}
```
```{r}
# Summary table metrics
dat_agg %>% group_by(email_group) %>% email_summary_table()
```

Drilling down into this None group:

```{r}
dat_agg %>%
  filter(email_group == 'None') %>%
  group_by(FIRST_MASTERS_YEAR, PROGRAM_GROUP) %>%
  summarise(n = length(ID_NUMBER))
```

Adding these groups to the above table yields the following.

```{r}
dat_agg <- dat_agg %>% mutate(
  email_group_filled = case_when(
    as.character(email_group) != 'None' ~ as.character(email_group)
    , PROGRAM_GROUP == 'EMBA' ~ 'Financial Times 2016 EMBA'
    , FIRST_MASTERS_YEAR == 2013 ~ 'Businessweek 2013 FT'
    , FIRST_MASTERS_YEAR == 2016 ~ 'Financial Times 2016 FT'
    , FIRST_MASTERS_YEAR == 2018 ~ 'Economist 2018 FT'
    , TRUE ~ 'None'
  ) %>% as.factor()
  , 
  # Add a version of email_group_filled with line breaks
  email_group_filled_lb = email_group_filled %>%
    fct_recode(
      'Financial Times\n2016 EMBA' = 'Financial Times 2016 EMBA'
      , 'Businessweek\n2013 FT' = 'Businessweek 2013 FT'
      , 'Financial Times\n2016 FT' = 'Financial Times 2016 FT'
      , 'Economist\n2018 FT' = 'Economist 2018 FT'
    )
)
```
```{r}
dat_agg %>% group_by(email_group_filled) %>% email_summary_table()
```

It's possible that link tracking was not enabled for Financial Times and the captured clicks were to non-survey links.

# Email address analysis

Just because an email was received doesn't mean it went to the individual's inbox (spam filters) or or even to an account still being checked (old address). I'd like to look at opens as a function of email address status (per the database) and email update date.

First, is the email sent to in iModules the same as the person's current preferred email address in the CATracks database?

```{r}
dat_agg %>% group_by(email_group_filled, same_email) %>% email_summary_table()
```
```{r}
# Function to create nice CI plots
plot_cis <- function(
  data, xdat, ydat, ndat, colordat, ci_mult = 1.96
  , title = '', xlab = '', ylab = '', collab = '', pd = .8, ci_width = .1
) {
  # Symbolic representation of data columns
  xdat <- ensym(xdat)
  ydat <- ensym(ydat)
  ndat <- ensym(ndat)
  colordat <- ensym(colordat)
  # Create data frame
  data %>% 
    group_by(!!xdat, !!colordat) %>%
    email_summary_table(FALSE) %>%
    mutate(
      sem = sqrt(!!ydat * (1 - !!ydat) / !!ndat)
      , ci = sem * ci_mult
    ) %>%
    # Plot results
    ggplot(aes(x = !!xdat, y = !!ydat, color = !!colordat)) +
    geom_point(position = position_dodge(pd)) +
    geom_errorbar(aes(ymin = !!ydat - ci, ymax = !!ydat + ci), position = position_dodge(pd), width = ci_width) +
    labs(x = xlab, y = ylab, color = collab, title = title) +
    scale_y_continuous(limits = c(0, 1), breaks = seq(0, 1, by = .2), labels = scales::percent)
}
```

## Open rates

```{r}
dat_agg %>% plot_cis(
  xdat = email_group_filled_lb, ydat = open_pct, ndat = sent, colordat = same_email
  , title = 'Open rate based on use of preferred email with 95% CI', ylab = 'Open rate', collab = 'Used pref email')
```

There appears to be a strong main effect for `same_email == TRUE`. Also note that all bounces, and not sent, emails are ones that weren't sent to the preferred email address per the database.

Other potentially important variables could be the email address status and date last updated.

```{r, rows.print = 100}
dat_agg %>% group_by(email_group_filled, send_email_status) %>% email_summary_table()
```
```{r, warning = FALSE}
dat_agg %>% plot_cis(
  xdat = email_group_filled_lb, ydat = open_pct, ndat = sent, colordat = send_email_status
  , title = 'Open rate based on use of send email status with 95% CI', ylab = 'Open rate', collab = 'Send email status')
```

Here, NA corresponds to emails not from the database. Again, note that these emails have the worst performance (with one exception due to expected small-sample variance).

Finally, consider open rate as a function of email age.

```{r, warning = FALSE, message = FALSE}
email_time_plot <- dat_agg %>%
  mutate(
    weeks_since_email = difftime(last_delivered_dt, STATUS_CHANGE_DATE, units = 'weeks') %>% as.numeric()
    , any_open = ifelse(n_opened_or_clicked > 0, 1, 0)
  ) %>%
  ggplot(aes(x = weeks_since_email, y = any_open)) +
  geom_point(aes(color = as.factor(any_open))) +
  geom_smooth(method = 'lm', color = 'darkgreen') +
  geom_smooth(method = 'loess') +
  scale_y_continuous(breaks = c(0, 1), minor_breaks = NULL) +
  labs(
    title = 'Email open rate as
    f(email delivery date, email address status update date)'
    , x = 'Weeks between email delivery date and last database preferred email status update'
    , y = 'Any email opened'
    , col = 'Any email opened'
  )
email_time_plot
```

This is actually pretty interesting -- there's an overall decreasing trend, with spikes around certain update times. I suspect a class year effect, so partitioning this out by email:

```{r, fig.height = 12, fig.width = 8, warning = FALSE, message = FALSE}
email_time_plot +
  facet_grid(email_group_filled_lb ~ .) +
  labs(title = '')
```

Interestingly, the peaks are within about the same time period, a little under 4 years before the last email was sent. Looking instead at the raw email update time data gives the following:

```{r, warning = FALSE, message = FALSE}
email_year_plot <- dat_agg %>%
  mutate(
    any_open = ifelse(n_opened_or_clicked > 0, 1, 0)
    , STATUS_CHANGE_DATE = as.Date(STATUS_CHANGE_DATE)
  ) %>%
  ggplot(aes(x = STATUS_CHANGE_DATE, y = any_open)) +
  geom_point(aes(color = as.factor(any_open))) +
  geom_smooth(method = 'lm', color = 'darkgreen') +
  geom_smooth(method = 'loess') +
  scale_x_date(date_breaks = '1 year', date_labels = '%Y') +
  scale_y_continuous(breaks = c(0, 1), minor_breaks = NULL) +
  labs(
    title = 'Email open rate as
    f(email address status update date)'
    , x = 'Date of last preferred email update'
    , y = 'Any email opened'
    , col = 'Any email opened'
  ) +
  theme(panel.grid.minor = element_line(linetype = 'dotted'))
email_year_plot
```

```{r, fig.height = 12, fig.width = 8, warning = FALSE, message = FALSE}
email_year_plot +
  facet_grid(email_group_filled_lb ~ .) +
  labs(title = '')
```

So around 4 years ago from summer 2015 to summer 2016 a bunch of good emails were captured, even from the class of 2013? Peculiar.

Plugging this all into a quick and simple model:

```{r}
dat_agg <- dat_agg %>%
  mutate(
    any_open = ifelse(n_opened_or_clicked > 0, 1, 0)
    , any_click = ifelse(n_clicks > 0, 1, 0)
    , weeks_since_email = difftime(last_delivered_dt, STATUS_CHANGE_DATE, units = 'weeks') %>% as.numeric()
    , send_email_status = case_when(
      is.na(send_email_status) ~ 'NA'
      , TRUE ~ as.character(send_email_status)
    ) %>% as.factor()
  )

glm_opens <- dat_agg %>%
  glm(
    any_open ~ email_group_filled + same_email + send_email_status + ns(weeks_since_email, df = 4)
    , data = .
    , family = binomial()
  )
```
```{r}
summary(glm_opens)
```

Note that the reference factor levels were Businessweek; sending emails to a non-preferred address; and email status active.

```{r}
crPlots(glm_opens, terms = ~ email_group_filled)
crPlots(glm_opens, terms = ~ same_email)
crPlots(glm_opens, terms = ~ send_email_status)
crPlots(glm_opens, terms = ~ ns(weeks_since_email, df = 4))
```

## Clicks

Now consider the same analyses with click data.

```{r, warning = FALSE}
dat_agg %>%
  plot_cis(
    xdat = email_group_filled_lb, ydat = clicked_pct, ndat = opened, colordat = same_email
    , title = 'Click rate based on use of preferred email with 95% CI', ylab = 'Click rate', collab = 'Used pref email'
  )
```

Note that the click rates are very low for the Financial Times surveys; again, I wonder whether the data was not captured correctly. These two should probably be modeled separately from the others.

```{r, warning = FALSE}
dat_agg %>% plot_cis(
  xdat = email_group_filled_lb, ydat = clicked_pct, ndat = opened, colordat = send_email_status
  , title = 'Click rate based on use of send email status with 95% CI', ylab = 'Open rate', collab = 'Send email status')
```

```{r, warning = FALSE, message = FALSE}
click_time_plot <- dat_agg %>%
  mutate(
    weeks_since_email = difftime(last_delivered_dt, STATUS_CHANGE_DATE, units = 'weeks') %>% as.numeric()
    , any_open = ifelse(n_opened_or_clicked > 0, 1, 0)
  ) %>%
  ggplot(aes(x = weeks_since_email, y = any_click)) +
  geom_point(aes(color = as.factor(any_click))) +
  geom_smooth(method = 'lm', color = 'darkgreen') +
  geom_smooth(method = 'loess') +
  scale_y_continuous(breaks = c(0, 1), minor_breaks = NULL) +
  labs(
    title = 'Email click rate as
    f(email delivery date, email address status update date)'
    , x = 'Weeks between email delivery date and last database preferred email status update'
    , y = 'Any email opened'
    , col = 'Any email opened'
  )
click_time_plot
```

Looks like the same trend as for opens, just the click rate starts much lower. Ignoring the two Financial Times surveys:

```{r, warning = FALSE, message = FALSE}
click_time_plot_filt <- dat_agg %>%
  mutate(
    weeks_since_email = difftime(last_delivered_dt, STATUS_CHANGE_DATE, units = 'weeks') %>% as.numeric()
    , any_open = ifelse(n_opened_or_clicked > 0, 1, 0)
  ) %>%
  filter(email_group_filled %in% c('Businessweek 2013 FT', 'Economist 2018 FT')) %>%
  ggplot(aes(x = weeks_since_email, y = any_click)) +
  geom_point(aes(color = as.factor(any_click))) +
  geom_smooth(method = 'lm', color = 'darkgreen') +
  geom_smooth(method = 'loess') +
  scale_y_continuous(breaks = c(0, 1), minor_breaks = NULL) +
  labs(
    title = 'Email click rate as
    f(email delivery date, email address status update date)'
    , x = 'Weeks between email delivery date and last database preferred email status update'
    , y = 'Any email opened'
    , col = 'Any email opened'
  )
click_time_plot_filt
```

So the noisiness was pretty much all due to Financial Times. Now breaking things out by publication:

```{r, fig.height = 12, fig.width = 8, warning = FALSE, message = FALSE}
click_time_plot +
  facet_grid(email_group_filled_lb ~ .) +
  labs(title = '')
```

And finally trying a quick model:

```{r}
glm_clicks <- dat_agg %>%
  filter(email_group_filled %in% c('Businessweek 2013 FT', 'Economist 2018 FT')) %>%
  glm(
    any_open ~ email_group_filled + same_email + send_email_status + ns(weeks_since_email, df = 4)
    , data = .
    , family = binomial()
  )
```
```{r}
summary(glm_clicks)
```

So email status doesn't provide any information beyond knowing that the preferred email was used.

```{r}
glm_clicks2 <- dat_agg %>%
  filter(email_group_filled %in% c('Businessweek 2013 FT', 'Economist 2018 FT')) %>%
  glm(
    any_open ~ email_group_filled + same_email + ns(weeks_since_email, df = 4)
    , data = .
    , family = binomial()
  )
```
```{r}
summary(glm_clicks2)
```

```{r}
crPlots(glm_clicks2, terms = ~ email_group_filled)
crPlots(glm_clicks2, terms = ~ same_email)
crPlots(glm_clicks2, terms = ~ ns(weeks_since_email, df = 4))
```

```{r}
# Save current data
save(dat_agg, email_summary_table, plot_cis, file = 'data/dat_agg2.Rdata')
```
